# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
...
		"IUser": {
			"properties": {
				"firstName": {
					"type": "string",
					"x-nullable": true
				},
				"lastName": {
					"type": "string",
					"x-nullable": true
				},
				"sex": {
					"type": "string",
					"enum": [
						"male",
						"female"
					],
					"x-nullable": true
				},
				"mobile": {
					"type": "string",
					"x-nullable": true
				},
				"expireDate": {
					"type": "string",
					"format": "date-time",
					"x-nullable": true
				},
				"registerDate": {
					"type": "string",
					"format": "date-time"
				}
			},
			"required": [
				"registerDate"
			],
			"type": "object",
			"additionalProperties": false
		},
...
	},
	"info": {
		"title": "app-platform-be",
		"version": "1.0.0",
		"description": "app platform backend",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "List",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/IUser"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUser"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUser"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "Find",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUser"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "body",
						"name": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUser"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}